Partition `io` module
