+++ src/io/device.rs
@@ -5,6 +5,7 @@ use std::fmt::Formatter;
 
 use crate::errors::Result;
 use crate::io;
+use crate::io::sensor::Sensor;
 use crate::storage::{Container, Containerized};
 
 /// Basic interface for GPIO device metadata
@@ -12,44 +13,6 @@ pub trait Device {
     fn get_metadata(&self) -> &DeviceMetadata;
     fn name(&self) -> String;
     fn id(&self) -> i32;
-}
-
-/// Interface for an input device
-/// It is used as a trait object and can be stored in a container using the `Containerized` trait.
-///
-/// # Type Parameters
-/// - `T`: The type of data that the sensor will produce.
-///
-/// # Functions
-/// - `read() -> T`: Reads the sensor and return the data as a value of type `T`.
-/// - `get_event() -> &IOEvent`: Create an `IOEvent` with current sensor data.
-///
-/// # Examples
-/// ```
-/// let sensor: &dyn crate::Sensor<f32> = TemperatureSensor::new(String::from("Temperature Sensor"), 1);
-/// let reading = sensor.read();
-/// let info = sensor.get_info();
-/// ```
-///
-/// # Notes:
-/// Since `Containerized` is implemented for the `Sensor` trait, any types that implement the `Sensor` trait
-/// can be stored in a container using the `Containerized::container()` method. This way, multiple instances of
-/// differing types may be stored in the same `Container`.
-///
-/// ```
-/// let mut container = Containerized::<Box<dyn crate::Sensor<f32>>, i32>::container();
-/// container.insert(1, Box::new(TemperatureSensor::new(String::from("Temperature Sensor"), 1)));
-/// container.insert(2, Box::new(HumiditySensor::new(String::from("Humidity Sensor"), 2)));
-/// ```
-/// > Note how two different sensor types were stored in `container`.
-pub trait Sensor: Device {
-    fn read(&self) -> f64;
-
-    fn get_event(&self, dt: DateTime<Utc>) -> io::IOEvent {
-        io::IOEvent::create(self, dt, self.read())
-    }
-
-    fn poll(&mut self, time: DateTime<Utc>) -> Result<()>;
 }
 
 impl std::fmt::Debug for dyn Sensor {
@@ -135,16 +98,5 @@ impl std::fmt::Display for DeviceMetadata {
             "Device Info {{ Kind: {} }}",
             self.kind,
         )
-    }
-}
-
-/// Returns a new instance of `Container` for storing objects which implement the `Sensor` trait which are accessed ``
-/// Objects are stored as `Box<dyn Sensor>`
-impl<K> Containerized<Box<dyn Sensor>, K> for dyn Sensor
-where
-    K: std::hash::Hash + Eq,
-{
-    fn container() -> Container<Box<dyn Sensor>, K> {
-        Container::<Box<dyn Sensor>, K>::new()
     }
 }
