+++ src/lib.rs
@@ -1,10 +1,8 @@
 extern crate chrono;
 
 pub mod errors;
-pub mod device;
 pub mod io;
 pub mod polling;
-pub mod sensors;
 pub mod settings;
 pub mod storage;
 pub mod units;
